rename semantikon to flowrep
